Mod/Robot: Removed -Wsign-compare warnings.

+++ src/Mod/Robot/App/kdl_cp/chainfksolvervel_recursive.cpp
@@ -43,11 +43,11 @@ namespace KDL
 
         if(!(in.q.rows()==chain.getNrOfJoints()&&in.qdot.rows()==chain.getNrOfJoints()))
             return -1;
-        else if(segmentNr>chain.getNrOfSegments())
+        else if(segmentNr>static_cast<int>(chain.getNrOfSegments()))
             return -1;
         else{
             int j=0;
-            for (unsigned int i=0;i<segmentNr;i++) {
+            for (int i=0;i<segmentNr;i++) {
                 //Calculate new Frame_base_ee
                 if(chain.getSegment(i).getJoint().getType()!=Joint::None){
                     out=out*FrameVel(chain.getSegment(i).pose(in.q(j)),
